rename to rime
