+++ docs/source/conf.py
@@ -17,7 +17,7 @@
 
 # -- Project information -----------------------------------------------------
 
-project = 'Recurrent Intensity Model Experiments (RIME)'
+project = 'Recurrent Intensity Model Experiments'
 copyright = '2021, Yifei Ma, Ge Liu, Anoop Deoras'
 author = 'Yifei Ma, Ge Liu, Anoop Deoras'
