+++ src/rime/models/hawkes.py
