+++ src/rime/models/hawkes_poisson.py
