+++ src/rime/models/implicit.py
