+++ src/rime/util/cvx_bisect.py
