+++ src/rime/util/score_array.py
